Prevent unintended parse
The termcap_path subroutine uses split with a capture on the colon or space
delimited files in `$ENV{TERMPATH}`.

This includes the separator in the split output. These elements are usually
removed by the last line of the subroutine:

    return grep { defined $_ && -f $_ } @termcap_path;

However if a colon-delimited TERMPATH was provided and the file ":" exists,
Term::Cap will attempt to parse it.

This (low) risk is prevented, and the code made more efficient by not
using a capture in the split.

Thanks